++ New PasswordCracker File
The file uploaded contains the code base for the Iterative password cracker I recently built. The cracker incorporates the password generator I built and iteration to generate strings of specified lengths. Theoretically, this can be used as a BruteForce password Cracker, although there are so many unique generations that it really isn't efficient or scalable. (In the case you're actually serious about cracking a password)